feature: convert more CommandArguments to ArgumentParser
