+++ cloud-minecraft/cloud-bungee/src/main/java/cloud/commandframework/bungee/BungeeCommandManager.java
@@ -25,8 +25,8 @@ package cloud.commandframework.bungee;
 
 import cloud.commandframework.CommandManager;
 import cloud.commandframework.CommandTree;
-import cloud.commandframework.bungee.arguments.PlayerArgument;
-import cloud.commandframework.bungee.arguments.ServerArgument;
+import cloud.commandframework.bungee.arguments.PlayerParser;
+import cloud.commandframework.bungee.arguments.ServerParser;
 import cloud.commandframework.captions.FactoryDelegatingCaptionRegistry;
 import cloud.commandframework.execution.CommandExecutionCoordinator;
 import cloud.commandframework.execution.FilteringCommandSuggestionProcessor;
@@ -86,9 +86,9 @@ public class BungeeCommandManager<C> extends CommandManager<C> {
 
         /* Register Bungee Parsers */
         this.parserRegistry().registerParserSupplier(TypeToken.get(ProxiedPlayer.class), parserParameters ->
-                new PlayerArgument.PlayerParser<>());
+                new PlayerParser<>());
         this.parserRegistry().registerParserSupplier(TypeToken.get(ServerInfo.class), parserParameters ->
-                new ServerArgument.ServerParser<>());
+                new ServerParser<>());
 
         /* Register default captions */
         if (this.captionRegistry() instanceof FactoryDelegatingCaptionRegistry) {
