+++ cloud-minecraft/cloud-bungee/src/main/java/cloud/commandframework/bungee/arguments/ServerParser.java
@@ -1,0 +1,113 @@
+//
+// MIT License
+//
+// Copyright (c) 2022 Alexander Söderberg & Contributors
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+package cloud.commandframework.bungee.arguments;
+
+import cloud.commandframework.CommandComponent;
+import cloud.commandframework.arguments.parser.ArgumentParseResult;
+import cloud.commandframework.arguments.parser.ArgumentParser;
+import cloud.commandframework.arguments.parser.ParserDescriptor;
+import cloud.commandframework.bungee.BungeeCaptionKeys;
+import cloud.commandframework.captions.CaptionVariable;
+import cloud.commandframework.context.CommandContext;
+import cloud.commandframework.context.CommandInput;
+import cloud.commandframework.exceptions.parsing.NoInputProvidedException;
+import cloud.commandframework.exceptions.parsing.ParserException;
+import net.md_5.bungee.api.ProxyServer;
+import net.md_5.bungee.api.config.ServerInfo;
+import org.apiguardian.api.API;
+import org.checkerframework.checker.nullness.qual.NonNull;
+
+/**
+ * Argument parser for {@link net.md_5.bungee.api.config.ServerInfo servers}
+ *
+ * @param <C> Command sender type
+ * @since 2.0.0
+ */
+public final class ServerParser<C> implements ArgumentParser<C, ServerInfo> {
+
+    /**
+     * Creates a new server parser.
+     *
+     * @param <C> command sender type
+     * @return the created parser
+     * @since 2.0.0
+     */
+    @API(status = API.Status.STABLE, since = "2.0.0")
+    public static <C> @NonNull ParserDescriptor<C, ServerInfo> serverParser() {
+        return ParserDescriptor.of(new ServerParser<>(), ServerInfo.class);
+    }
+
+    /**
+     * Returns a {@link CommandComponent.Builder} using {@link #serverParser()} as the parser.
+     *
+     * @param <C> the command sender type
+     * @return the component builder
+     * @since 2.0.0
+     */
+    @API(status = API.Status.STABLE, since = "2.0.0")
+    public static <C> CommandComponent.@NonNull Builder<C, ServerInfo> serverComponent() {
+        return CommandComponent.<C, ServerInfo>builder().parser(serverParser());
+    }
+
+    @Override
+    public @NonNull ArgumentParseResult<@NonNull ServerInfo> parse(
+            final @NonNull CommandContext<@NonNull C> commandContext,
+            final @NonNull CommandInput commandInput
+    ) {
+        final String input = commandInput.peekString();
+        if (input.isEmpty()) {
+            return ArgumentParseResult.failure(new NoInputProvidedException(
+                    ServerParser.class,
+                    commandContext
+            ));
+        }
+        final ServerInfo server = commandContext.<ProxyServer>get("ProxyServer").getServerInfo(commandInput.readString());
+        if (server == null) {
+            return ArgumentParseResult.failure(
+                    new ServerParseException(
+                            input,
+                            commandContext
+                    )
+            );
+        }
+        return ArgumentParseResult.success(server);
+    }
+
+    public static final class ServerParseException extends ParserException {
+
+        private static final long serialVersionUID = -3825941611365494659L;
+
+        private ServerParseException(
+                final @NonNull String input,
+                final @NonNull CommandContext<?> context
+        ) {
+            super(
+                    ServerParser.class,
+                    context,
+                    BungeeCaptionKeys.ARGUMENT_PARSE_FAILURE_SERVER,
+                    CaptionVariable.of("input", input)
+            );
+        }
+    }
+}
