+++ cloud-minecraft/cloud-velocity/src/main/java/cloud/commandframework/velocity/VelocityCommandManager.java
@@ -32,8 +32,8 @@ import cloud.commandframework.execution.CommandExecutionCoordinator;
 import cloud.commandframework.execution.FilteringCommandSuggestionProcessor;
 import cloud.commandframework.meta.CommandMeta;
 import cloud.commandframework.meta.SimpleCommandMeta;
-import cloud.commandframework.velocity.arguments.PlayerArgument;
-import cloud.commandframework.velocity.arguments.ServerArgument;
+import cloud.commandframework.velocity.arguments.PlayerParser;
+import cloud.commandframework.velocity.arguments.ServerParser;
 import com.google.inject.Inject;
 import com.google.inject.Module;
 import com.google.inject.Singleton;
@@ -105,9 +105,9 @@ public class VelocityCommandManager<C> extends CommandManager<C> implements Brig
 
         /* Register Velocity Parsers */
         this.parserRegistry().registerParserSupplier(TypeToken.get(Player.class), parserParameters ->
-                new PlayerArgument.PlayerParser<>());
+                new PlayerParser<>());
         this.parserRegistry().registerParserSupplier(TypeToken.get(RegisteredServer.class), parserParameters ->
-                new ServerArgument.ServerParser<>());
+                new ServerParser<>());
 
         /* Register default captions */
         if (this.captionRegistry() instanceof FactoryDelegatingCaptionRegistry) {
