+++ cloud-minecraft/cloud-velocity/src/main/java/cloud/commandframework/velocity/arguments/PlayerParser.java
@@ -1,0 +1,127 @@
+//
+// MIT License
+//
+// Copyright (c) 2022 Alexander Söderberg & Contributors
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+package cloud.commandframework.velocity.arguments;
+
+import cloud.commandframework.CommandComponent;
+import cloud.commandframework.arguments.parser.ArgumentParseResult;
+import cloud.commandframework.arguments.parser.ArgumentParser;
+import cloud.commandframework.arguments.parser.ParserDescriptor;
+import cloud.commandframework.captions.CaptionVariable;
+import cloud.commandframework.context.CommandContext;
+import cloud.commandframework.context.CommandInput;
+import cloud.commandframework.exceptions.parsing.NoInputProvidedException;
+import cloud.commandframework.exceptions.parsing.ParserException;
+import cloud.commandframework.velocity.VelocityCaptionKeys;
+import com.velocitypowered.api.proxy.Player;
+import com.velocitypowered.api.proxy.ProxyServer;
+import com.velocitypowered.api.proxy.server.RegisteredServer;
+import java.util.List;
+import java.util.stream.Collectors;
+import org.apiguardian.api.API;
+import org.checkerframework.checker.nullness.qual.NonNull;
+
+/**
+ * Argument parser for {@link Player players}
+ *
+ * @param <C> Command sender type
+ * @since 2.0.0
+ */
+public final class PlayerParser<C> implements ArgumentParser<C, Player> {
+
+    /**
+     * Creates a new player parser.
+     *
+     * @param <C> command sender type
+     * @return the created parser
+     * @since 2.0.0
+     */
+    @API(status = API.Status.STABLE, since = "2.0.0")
+    public static <C> @NonNull ParserDescriptor<C, RegisteredServer> playerParser() {
+        return ParserDescriptor.of(new ServerParser<>(), RegisteredServer.class);
+    }
+
+    /**
+     * Returns a {@link CommandComponent.Builder} using {@link #playerParser()} as the parser.
+     *
+     * @param <C> the command sender type
+     * @return the component builder
+     * @since 2.0.0
+     */
+    @API(status = API.Status.STABLE, since = "2.0.0")
+    public static <C> CommandComponent.@NonNull Builder<C, RegisteredServer> playerComponent() {
+        return CommandComponent.<C, RegisteredServer>builder().parser(playerParser());
+    }
+
+    @Override
+    public @NonNull ArgumentParseResult<@NonNull Player> parse(
+            final @NonNull CommandContext<@NonNull C> commandContext,
+            final @NonNull CommandInput commandInput
+    ) {
+        final String input = commandInput.peekString();
+        if (input.isEmpty()) {
+            return ArgumentParseResult.failure(new NoInputProvidedException(
+                    PlayerParser.class,
+                    commandContext
+            ));
+        }
+        final Player player = commandContext.<ProxyServer>get("ProxyServer")
+                .getPlayer(commandInput.readString())
+                .orElse(null);
+        if (player == null) {
+            return ArgumentParseResult.failure(
+                    new PlayerParseException(
+                            input,
+                            commandContext
+                    )
+            );
+        }
+        return ArgumentParseResult.success(player);
+    }
+
+    @Override
+    public @NonNull List<@NonNull String> stringSuggestions(
+            final @NonNull CommandContext<C> commandContext,
+            final @NonNull String input
+    ) {
+        return commandContext.<ProxyServer>get("ProxyServer").getAllPlayers()
+                .stream().map(Player::getUsername).collect(Collectors.toList());
+    }
+
+    public static final class PlayerParseException extends ParserException {
+
+        private static final long serialVersionUID = -4839583631837040297L;
+
+        private PlayerParseException(
+                final @NonNull String input,
+                final @NonNull CommandContext<?> context
+        ) {
+            super(
+                    PlayerParser.class,
+                    context,
+                    VelocityCaptionKeys.ARGUMENT_PARSE_FAILURE_PLAYER,
+                    CaptionVariable.of("input", input)
+            );
+        }
+    }
+}
