+++ cloud-minecraft/cloud-velocity/src/main/java/cloud/commandframework/velocity/arguments/ServerParser.java
@@ -1,0 +1,129 @@
+//
+// MIT License
+//
+// Copyright (c) 2022 Alexander Söderberg & Contributors
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+package cloud.commandframework.velocity.arguments;
+
+import cloud.commandframework.CommandComponent;
+import cloud.commandframework.arguments.parser.ArgumentParseResult;
+import cloud.commandframework.arguments.parser.ArgumentParser;
+import cloud.commandframework.arguments.parser.ParserDescriptor;
+import cloud.commandframework.captions.CaptionVariable;
+import cloud.commandframework.context.CommandContext;
+import cloud.commandframework.context.CommandInput;
+import cloud.commandframework.exceptions.parsing.NoInputProvidedException;
+import cloud.commandframework.exceptions.parsing.ParserException;
+import cloud.commandframework.velocity.VelocityCaptionKeys;
+import com.velocitypowered.api.proxy.ProxyServer;
+import com.velocitypowered.api.proxy.server.RegisteredServer;
+import java.util.List;
+import java.util.stream.Collectors;
+import org.apiguardian.api.API;
+import org.checkerframework.checker.nullness.qual.NonNull;
+
+/**
+ * Argument parser for {@link RegisteredServer servers}
+ *
+ * @param <C> Command sender type
+ * @since 2.0.0
+ */
+public final class ServerParser<C> implements ArgumentParser<C, RegisteredServer> {
+
+    /**
+     * Creates a new server parser.
+     *
+     * @param <C> command sender type
+     * @return the created parser
+     * @since 2.0.0
+     */
+    @API(status = API.Status.STABLE, since = "2.0.0")
+    public static <C> @NonNull ParserDescriptor<C, RegisteredServer> serverParser() {
+        return ParserDescriptor.of(new ServerParser<>(), RegisteredServer.class);
+    }
+
+    /**
+     * Returns a {@link CommandComponent.Builder} using {@link #serverParser()} as the parser.
+     *
+     * @param <C> the command sender type
+     * @return the component builder
+     * @since 2.0.0
+     */
+    @API(status = API.Status.STABLE, since = "2.0.0")
+    public static <C> CommandComponent.@NonNull Builder<C, RegisteredServer> serverComponent() {
+        return CommandComponent.<C, RegisteredServer>builder().parser(serverParser());
+    }
+
+    @Override
+    public @NonNull ArgumentParseResult<@NonNull RegisteredServer> parse(
+            final @NonNull CommandContext<@NonNull C> commandContext,
+            final @NonNull CommandInput commandInput
+    ) {
+        final String input = commandInput.peekString();
+        if (input.isEmpty()) {
+            return ArgumentParseResult.failure(new NoInputProvidedException(
+                    ServerParser.class,
+                    commandContext
+            ));
+        }
+        final RegisteredServer server = commandContext.<ProxyServer>get("ProxyServer")
+                .getServer(commandInput.readString())
+                .orElse(null);
+        if (server == null) {
+            return ArgumentParseResult.failure(
+                    new ServerParseException(
+                            input,
+                            commandContext
+                    )
+            );
+        }
+        return ArgumentParseResult.success(server);
+    }
+
+    @Override
+    public @NonNull List<@NonNull String> stringSuggestions(
+            final @NonNull CommandContext<C> commandContext,
+            final @NonNull String input
+    ) {
+        return commandContext.<ProxyServer>get("ProxyServer")
+                .getAllServers()
+                .stream()
+                .map(s -> s.getServerInfo().getName())
+                .collect(Collectors.toList());
+    }
+
+    public static final class ServerParseException extends ParserException {
+
+        private static final long serialVersionUID = 9168156226853233788L;
+
+        private ServerParseException(
+                final @NonNull String input,
+                final @NonNull CommandContext<?> context
+        ) {
+            super(
+                    ServerParser.class,
+                    context,
+                    VelocityCaptionKeys.ARGUMENT_PARSE_FAILURE_SERVER,
+                    CaptionVariable.of("input", input)
+            );
+        }
+    }
+}
