+++ examples/example-velocity/src/main/java/cloud/commandframework/examples/velocity/ExampleVelocityPlugin.java
@@ -28,8 +28,8 @@ import cloud.commandframework.minecraft.extras.AudienceProvider;
 import cloud.commandframework.minecraft.extras.MinecraftExceptionHandler;
 import cloud.commandframework.velocity.CloudInjectionModule;
 import cloud.commandframework.velocity.VelocityCommandManager;
-import cloud.commandframework.velocity.arguments.PlayerArgument;
-import cloud.commandframework.velocity.arguments.ServerArgument;
+import cloud.commandframework.velocity.arguments.PlayerParser;
+import cloud.commandframework.velocity.arguments.ServerParser;
 import com.google.inject.Inject;
 import com.google.inject.Injector;
 import com.google.inject.Key;
@@ -88,7 +88,7 @@ public final class ExampleVelocityPlugin {
                 .apply(commandManager, AudienceProvider.nativeAudience());
         commandManager.command(
                 commandManager.commandBuilder("example")
-                        .required(PlayerArgument.of("player"))
+                        .required("player", PlayerParser.playerParser())
                         .handler(context -> {
                                     final Player player = context.get("player");
                                     context.getSender().sendMessage(
@@ -104,7 +104,7 @@ public final class ExampleVelocityPlugin {
         );
         commandManager.command(
                 commandManager.commandBuilder("example-server")
-                        .required(ServerArgument.of("server"))
+                        .required("server", ServerParser.serverParser())
                         .handler(context -> {
                             final RegisteredServer server = context.get("server");
                             context.getSender().sendMessage(
